refactor: move ValidationState to its own file
